Update test endpoint that is missing path
amqp-common 0.1.4 and later introduced a different error for the
invalid endpoint of "abc". This change update the endpoint so that we
can still test the missing path error, which technically is a
different error thrown from a different method starting from
amqp-common 0.1.4 or later. It just have the same error type and error
message as the old one.